+++ Game/Scripts/LightFinisherAttackScript.cpp
