+++ Source/Engine.vcxproj
@@ -697,6 +697,7 @@
     <ClCompile Include="FileSystem\ModuleResources.cpp" />
     <ClCompile Include="DataModels\GBuffer\GBuffer.cpp" />
     <ClCompile Include="FileSystem\UIDGenerator.cpp" />
+    <ClCompile Include="LightAttackBullet.cpp" />
     <ClCompile Include="LightFinisherAttackScript.cpp" />
     <ClCompile Include="Main.cpp" />
     <ClCompile Include="DataModels\ParticleSystem\ModuleRenderer.cpp" />
@@ -1157,6 +1158,7 @@
     <ClInclude Include="FileSystem\ModuleResources.h" />
     <ClInclude Include="DataModels\GBuffer\GBuffer.h" />
     <ClInclude Include="DataModels\ParticleSystem\ModuleRenderer.h" />
+    <ClInclude Include="LightAttackBullet.h" />
     <ClInclude Include="LightFinisherAttackScript.h" />
     <ClInclude Include="Modules\ModuleNavigation.h" />
     <ClInclude Include="FileSystem\UID.h" />
