+++ Source/Engine.vcxproj.filters
@@ -616,6 +616,7 @@
       <Filter>Auxiliar</Filter>
     </ClCompile>
     <ClCompile Include="LightFinisherAttackScript.cpp" />
+    <ClCompile Include="LightAttackBullet.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="DataModels\Timer\Timer.h">
@@ -1277,6 +1278,7 @@
     </ClInclude>
     <ClInclude Include="..\Game\Scripts\DefaultScript.h" />
     <ClInclude Include="LightFinisherAttackScript.h" />
+    <ClInclude Include="LightAttackBullet.h" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="Auxiliar">
